tests(cksum): Regroup common tests under a single macro
